refactor: move usageHistory service to history/usageHistory directory
Update all imports to reflect new service location and maintain
consistent path structure across the codebase. This change
organizes related services under a common history namespace
and improves project structure clarity.
