+++ entrypoints/background/runtimeMessages.ts
@@ -11,10 +11,10 @@ import { handleChannelConfigMessage } from "~/services/channelConfigStorage"
 import { handleAutoCheckinMessage } from "~/services/checkin/autoCheckin/scheduler"
 import { handleExternalCheckInMessage } from "~/services/checkin/externalCheckInService"
 import { handleDailyBalanceHistoryMessage } from "~/services/history/dailyBalanceHistory/scheduler"
+import { handleUsageHistoryMessage } from "~/services/history/usageHistory/scheduler"
 import { handleLdohSiteLookupMessage } from "~/services/integrations/ldohSiteLookup/background"
 import { handleManagedSiteModelSyncMessage } from "~/services/models/modelSync"
 import { handleRedemptionAssistMessage } from "~/services/redemptionAssist"
-import { handleUsageHistoryMessage } from "~/services/usageHistory/scheduler"
 import { handleWebAiApiCheckMessage } from "~/services/webAiApiCheck/background"
 import { handleWebdavAutoSyncMessage } from "~/services/webdav/webdavAutoSyncService"
 import { onRuntimeMessage } from "~/utils/browserApi"
