+++ entrypoints/background/servicesInit.ts
@@ -1,10 +1,10 @@
 import { autoRefreshService } from "~/services/accounts/autoRefreshService"
 import { autoCheckinScheduler } from "~/services/checkin/autoCheckin/scheduler"
 import { dailyBalanceHistoryScheduler } from "~/services/history/dailyBalanceHistory/scheduler"
+import { usageHistoryScheduler } from "~/services/history/usageHistory/scheduler"
 import { modelMetadataService } from "~/services/models/modelMetadata"
 import { modelSyncScheduler } from "~/services/models/modelSync"
 import { redemptionAssistService } from "~/services/redemptionAssist"
-import { usageHistoryScheduler } from "~/services/usageHistory/scheduler"
 import { webdavAutoSyncService } from "~/services/webdav/webdavAutoSyncService"
 import { initBackgroundI18n } from "~/utils/background-i18n"
 import { createLogger } from "~/utils/logger"
