+++ entrypoints/options/pages/BasicSettings/components/UsageHistorySyncTab.tsx
@@ -7,7 +7,7 @@ import { Card, CardContent, Input } from "~/components/ui"
 import { RuntimeActionIds } from "~/constants/runtimeActions"
 import { useUserPreferencesContext } from "~/contexts/UserPreferencesContext"
 import { accountStorage } from "~/services/accounts/accountStorage"
-import { usageHistoryStorage } from "~/services/usageHistory/storage"
+import { usageHistoryStorage } from "~/services/history/usageHistory/storage"
 import type { SiteAccount } from "~/types"
 import { USAGE_HISTORY_SCHEDULE_MODE } from "~/types/usageHistory"
 import type { UsageHistoryStore } from "~/types/usageHistory"
