+++ services/accounts/autoRefreshService.ts
@@ -1,10 +1,10 @@
 import { RuntimeActionIds } from "~/constants/runtimeActions"
+import { usageHistoryScheduler } from "~/services/history/usageHistory/scheduler"
 import { AccountAutoRefresh } from "~/types/accountAutoRefresh"
 import { sendRuntimeMessage } from "~/utils/browserApi"
 import { createLogger } from "~/utils/logger"
 
 import { getErrorMessage } from "../../utils/error"
-import { usageHistoryScheduler } from "../usageHistory/scheduler"
 import { userPreferences } from "../userPreferences"
 import { accountStorage } from "./accountStorage"
