+++ tests/entrypoints/options/UsageAnalytics.test.tsx
@@ -2,14 +2,14 @@ import { describe, expect, it, vi } from "vitest"
 
 import UsageAnalytics from "~/entrypoints/options/pages/UsageAnalytics"
 import { accountStorage } from "~/services/accounts/accountStorage"
-import { usageHistoryStorage } from "~/services/usageHistory/storage"
+import { usageHistoryStorage } from "~/services/history/usageHistory/storage"
 import { render, screen } from "~/tests/test-utils/render"
 
 vi.mock("~/services/accounts/accountStorage", () => ({
   accountStorage: { getAllAccounts: vi.fn() },
 }))
 
-vi.mock("~/services/usageHistory/storage", () => ({
+vi.mock("~/services/history/usageHistory/storage", () => ({
   usageHistoryStorage: { getStore: vi.fn() },
 }))
