+++ tests/entrypoints/options/pages/UsageAnalytics/UsageAnalyticsNavigation.test.tsx
@@ -3,8 +3,8 @@ import { describe, expect, it, vi } from "vitest"
 
 import UsageAnalytics from "~/entrypoints/options/pages/UsageAnalytics"
 import { accountStorage } from "~/services/accounts/accountStorage"
-import { createEmptyUsageHistoryAccountStore } from "~/services/usageHistory/core"
-import { usageHistoryStorage } from "~/services/usageHistory/storage"
+import { createEmptyUsageHistoryAccountStore } from "~/services/history/usageHistory/core"
+import { usageHistoryStorage } from "~/services/history/usageHistory/storage"
 import { render, screen } from "~/tests/test-utils/render"
 import { navigateWithinOptionsPage } from "~/utils/navigation"
 
@@ -29,7 +29,7 @@ vi.mock("~/services/accounts/accountStorage", () => ({
   accountStorage: { getAllAccounts: vi.fn() },
 }))
 
-vi.mock("~/services/usageHistory/storage", () => ({
+vi.mock("~/services/history/usageHistory/storage", () => ({
   usageHistoryStorage: { getStore: vi.fn() },
 }))
