+++ tests/services/usageHistory/sync.test.ts
@@ -3,9 +3,9 @@ import { describe, expect, it } from "vitest"
 
 import { accountStorage } from "~/services/accounts/accountStorage"
 import { LogType, type LogItem } from "~/services/apiService/common/type"
-import { getDayKeyFromUnixSeconds } from "~/services/usageHistory/core"
-import { usageHistoryStorage } from "~/services/usageHistory/storage"
-import { syncUsageHistoryForAccount } from "~/services/usageHistory/sync"
+import { getDayKeyFromUnixSeconds } from "~/services/history/usageHistory/core"
+import { usageHistoryStorage } from "~/services/history/usageHistory/storage"
+import { syncUsageHistoryForAccount } from "~/services/history/usageHistory/sync"
 import { server } from "~/tests/msw/server"
 import { AuthTypeEnum, SiteHealthStatus, type SiteAccount } from "~/types"
 import {
